+++ Lib.Models.TinyTransformer/State/TinyTransformerWeights.cs
@@ -5,28 +5,28 @@ namespace Lib.Models.TinyTransformer.State;
 public class TinyTransformerWeights
 {
     public static Random rnd = new Random(42);
-    public double[][] wQ { get; set; }
-    public double[][] wK { get; set; }
-    public double[][] wV { get; set; }
-    public double[][] wO { get; set; }
-    public double[][] OutputW { get; set; }
-    public double[][] ffn1 { get; set; }
-    public double[][] ffn2 { get; set; }
-    public double[] ffn1Bias { get; set; }
-    public double[] ffn2Bias { get; set; }
-    public double[] OutputBias { get; set; }
+    public float[][] wQ { get; set; }
+    public float[][] wK { get; set; }
+    public float[][] wV { get; set; }
+    public float[][] wO { get; set; }
+    public float[][] OutputW { get; set; }
+    public float[][] ffn1 { get; set; }
+    public float[][] ffn2 { get; set; }
+    public float[] ffn1Bias { get; set; }
+    public float[] ffn2Bias { get; set; }
+    public float[] OutputBias { get; set; }
 
-    public static double[][] GenerateMatrix(int rows, int cols)
+    public static float[][] GenerateMatrix(int rows, int cols)
     {
-        double[][] res = new double[rows][];
+        float[][] res = new float[rows][];
 
         for (int i = 0; i < rows; i++)
         {
-            res[i] = new double[cols];
+            res[i] = new float[cols];
 
             for (int j = 0; j < cols; j++)
             {
-                res[i][j] = rnd.NextDouble() * 0.2 - 0.1;
+                res[i][j] = rnd.NextSingle() * 0.2f - 0.1f;
             }
         }
 
@@ -48,8 +48,8 @@ public class TinyTransformerWeights
         ffn2 =
             GenerateMatrix(embeddingSize * 4, embeddingSize);
 
-        ffn1Bias = new double[embeddingSize * 4];
-        ffn2Bias = new double[embeddingSize];
-        OutputBias = new double[vocabSize];
+        ffn1Bias = new float[embeddingSize * 4];
+        ffn2Bias = new float[embeddingSize];
+        OutputBias = new float[vocabSize];
     }
 }
